Build custom utility for outputting markdown
I was hitting various minor issues when trying to use the
cooklang-cli or chef to build my markdown.  Increasingly it was clear
some of the things I wanted, like toml frontmatters with zola
taxonomies, just wouldn't fit in the upstream project.  This adds a
small rust utility that is based on cooklang-to-md.  It includes
specific formating that I care about and strips out some other logic.
This isn't meant to be reusable or useful to anyone but me.

A first really slim pass at taxonomies in here, but isn't usable as
is.